Add responsive sizing using clamp
To make this look better on mobile, I updated the CSS to use clamp to
calculate the font size and padding.

CSS clamp() sets a minimum size, a maximum size, and lets the browser
smoothly scale between them based on viewport width - eliminating the
need for breakpoints.

Using this article as a reference:
https://robmc.dev/blog/look-mum-no-breakpoints/

And this tool to do the calculations:
https://utopia.fyi/